+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 ### Added
  - Add juliet, a minimal minor variant caller
  - Add Fisher's exact test
- - Juliet JSON and HTMl output for HIV-1 gag and pol region
+ - Juliet JSON and HTMl output for HIV-1 pol region
+ - Call codon-wise
 
 ## [2.0.4]